css/css-paint-api/dynamic-import.https.html

<!DOCTYPE html>
<html class="reftest-wait">
<link rel="help" href="https://html.spec.whatwg.org/#hostimportmoduledynamically(referencingscriptormodule,-specifier,-promisecapability)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="output"></div>

<script id="code" type="text/worklet">
registerPaint('rects', class {
    async paint(ctx, geom) {
        await import('./resources/imported-module.mjs');
    }
});
</script>

<script>
"use strict";

async function construct_and_import_worklet() {
  const blob = new Blob([document.getElementById('code').textContent], {type: 'text/javascript'});
  let url = URL.createObjectURL(blob);

  await CSS.paintWorklet.addModule(url);

  return new Promise(requestAnimationFrame);
}

promise_test(t => {
  return promise_rejects_js(t, SyntaxError, construct_and_import_worklet());
}, "Dynamic import() inside a paintWorklet should throw an exception and fail.");
</script>
